put the lighting button in the toolbar
