+++ src/components/configure/configuration/ConfigurationDialog.tsx
@@ -24,7 +24,6 @@ import { Alert, AlertTitle } from '@material-ui/lab';
 import { KeyboardDefinitionSchema } from '../../../gen/types/KeyboardDefinition';
 import { KeyboardDefinitionFormPart } from '../../common/keyboarddefformpart/KeyboardDefinitionFormPart';
 import { hexadecimal } from '../../../utils/StringUtils';
-import TabLighting, { defaultUnderglowEffects, Hsv } from './TabLighting';
 
 const GOOGLE_FORM_URL =
   'https://docs.google.com/forms/d/e/1FAIpQLScZPhiXEG2VETCGZ2dYp4YbzzMlU62Crh1cNxPpFBkN4cCPbA/viewform?usp=pp_url&entry.661359702=${keyboard_name}&entry.135453541=${keyboard_id}';
@@ -45,28 +44,7 @@ type OwnState = {
   selectedMenuIndex: number;
   keyboardDefinition: KeyboardDefinitionSchema | null;
   keyboardDefinitionFile: string | null;
-  needToInit: boolean;
 };
-
-type LightingType =
-  | undefined
-  | (
-      | 'none'
-      | 'qmk_backlight'
-      | 'qmk_rgblight'
-      | 'qmk_backlight_rgblight'
-      | 'wt_rgb_backlight'
-      | 'wt_mono_backlight'
-    )
-  | {
-      extends?: string;
-      effects?: [] | [[] | [string] | [string, number]];
-      keycodes?: 'qmk' | 'wt';
-      supportedBacklightValues?: number[];
-      supportedLightingValues?: number[];
-      underglowEffects?: [] | [string] | [string, number][];
-      [k: string]: unknown;
-    };
 
 export default class ConfigurationDialog extends React.Component<
   ConfigurationDialogProps,
@@ -75,9 +53,6 @@ export default class ConfigurationDialog extends React.Component<
   private googleFormUrl: string = '';
   private githubUrl: string = '';
   private githubDisplayName: string = '';
-  private showUnderglow: boolean = true;
-  private showBacklight: boolean = true;
-  private underglowEffects: [string, number][] = [];
 
   constructor(
     props: ConfigurationDialogProps | Readonly<ConfigurationDialogProps>
@@ -87,68 +62,7 @@ export default class ConfigurationDialog extends React.Component<
       selectedMenuIndex: 0,
       keyboardDefinition: null,
       keyboardDefinitionFile: null,
-      needToInit: false,
     };
-    this.initLighting();
-  }
-
-  get showLighting() {
-    return this.showBacklight || this.showUnderglow;
-  }
-
-  shouldComponentUpdate(
-    // eslint-disable-next-line no-unused-vars
-    nextProps: ConfigurationDialogProps,
-    // eslint-disable-next-line no-unused-vars
-    nextState: OwnState
-  ) {
-    if (this.state.needToInit) {
-      this.initLighting();
-      this.setState({ needToInit: false });
-    }
-    return true;
-  }
-
-  private initLighting() {
-    const lighting: LightingType = this.props.keyboardDefinition?.lighting;
-    this.showUnderglow = false;
-    if (!lighting) {
-      this.showUnderglow = false;
-      this.showBacklight = false;
-      return;
-    }
-
-    if (typeof lighting === 'string') {
-      this.showUnderglow = 0 <= lighting.indexOf('rgblight');
-      this.showBacklight = 0 <= lighting.indexOf('backlight');
-      this.underglowEffects = defaultUnderglowEffects;
-      return;
-    }
-
-    if (!lighting.extends) {
-      /**
-       * lighting object MUST be contains 'extends' property.
-       * ref. https://caniusevia.com/docs/optional#lighting
-       */
-      throw new Error(
-        `lighting properties whose type is NOT 'string' MUST contain 'extends'.`
-      );
-    }
-    this.showUnderglow = 0 <= lighting.extends.indexOf('rgblight');
-    this.showBacklight = 0 <= lighting.extends.indexOf('backlight');
-
-    if (!lighting.underglowEffects || lighting.underglowEffects.length === 0) {
-      // use default effects if no overridden effects
-      this.underglowEffects = defaultUnderglowEffects;
-      return;
-    }
-
-    if (typeof lighting.underglowEffects[0] === 'string') {
-      const label: string = lighting.underglowEffects[0];
-      this.underglowEffects = [[label, 0]];
-    } else {
-      this.underglowEffects = lighting.underglowEffects as [string, number][];
-    }
   }
 
   private onEnter() {
@@ -171,42 +85,8 @@ export default class ConfigurationDialog extends React.Component<
   }
 
   private onClickApplyKeyboardDefinition() {
-    this.setState({ needToInit: true });
     this.props.refreshKeyboardDefinition!(this.state.keyboardDefinition!);
     this.clearKeyboardDefinition();
-  }
-
-  private onChangeBacklight(backlight: {
-    isBreathing?: boolean;
-    brightness?: number /* 0-100 */;
-  }) {
-    if (backlight.isBreathing != undefined) {
-      this.props.keyboard!.updateBacklightEffect(backlight.isBreathing);
-    }
-
-    if (backlight.brightness != undefined) {
-      const brightness = Math.round(255 * (backlight.brightness / 100));
-      this.props.keyboard!.updateBacklightBrightness(brightness);
-    }
-  }
-
-  private onChangeUnderglow(underglow: {
-    mode?: number;
-    color?: Hsv; // h: 0-360, s: 0-100, v: 0-100
-  }) {
-    if (underglow.mode != undefined) {
-      this.props.keyboard!.updateRGBLightEffect(underglow.mode);
-    }
-
-    if (underglow.color != undefined) {
-      const hsv: Hsv = underglow.color;
-      const hue = Math.round(255 * (hsv.h / 360));
-      const sat = Math.round(255 * (hsv.s / 100));
-      const brightness = Math.round(255 * (hsv.v / 100));
-
-      this.props.keyboard!.updateRGBLightColor(hue, sat);
-      this.props.keyboard!.updateRGBLightBrightness(brightness);
-    }
   }
 
   // eslint-disable-next-line no-unused-vars
@@ -256,7 +136,6 @@ export default class ConfigurationDialog extends React.Component<
             className="config-menu"
           >
             {hasKeyboardOptions && <Tab label="Layout options" />}
-            {this.showLighting && <Tab label="Lighting" />}
             <Tab label="Import" />
             <Tab label="Info" />
           </Tabs>
@@ -276,22 +155,6 @@ export default class ConfigurationDialog extends React.Component<
                   );
                 })}
               </Grid>
-            </TabPanel>
-          )}
-          {this.showLighting && (
-            <TabPanel value={this.state.selectedMenuIndex} index={panelIndex++}>
-              <TabLighting
-                underglowEffects={this.underglowEffects}
-                keyboard={this.props.keyboard!}
-                showBacklight={this.showBacklight}
-                showUnderglow={this.showUnderglow}
-                onChangeUnderglow={(underglow) => {
-                  this.onChangeUnderglow(underglow);
-                }}
-                onChangeBacklight={(backlight) => {
-                  this.onChangeBacklight(backlight);
-                }}
-              />
             </TabPanel>
           )}
           <TabPanel value={this.state.selectedMenuIndex} index={panelIndex++}>
