+++ src/components/configure/keymap/Keymap.scss
@@ -9,7 +9,6 @@
     display: flex;
     justify-content: flex-start;
     align-items: center;
-    padding: 8px 16px;
   }
 
   .label-lang,
