+++ src/components/configure/keymapToolbar/KeymapToolbar.scss
@@ -8,6 +8,6 @@
   align-items: flex-start;
   padding: 8px 16px;
   .keymap-menu-item {
-    margin-left: $space-m;
+    margin-bottom: $space-m;
   }
 }
