+++ src/components/lighting/Lighting.scss
@@ -1,4 +1,4 @@
-@import '../../../variables';
+@import '../../variables';
 
 .lighting-settings {
   h4 {
